requirement.txt
gunicorn
pillow
requests
torch
yolov5
clip-by-openai
opencv-python
flask
flask-cors
